+++ src/main/java/com/taja/bikeapi/application/dto/station/StationInfoDto.java
@@ -1,7 +1,7 @@
-package com.taja.collector.infra.api.dto.station;
+package com.taja.bikeapi.application.dto.station;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.taja.collector.infra.api.dto.status.ResultDto;
+import com.taja.bikeapi.application.dto.status.ResultDto;
 import java.util.List;
 
 public record StationInfoDto(
