+++ src/main/java/com/taja/bikeapi/application/dto/status/ResultDto.java
@@ -1,4 +1,4 @@
-package com.taja.collector.infra.api.dto.station;
+package com.taja.bikeapi.application.dto.status;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
