+++ src/main/java/com/taja/station/application/StationRedisRepository.java
@@ -1,6 +1,6 @@
 package com.taja.station.application;
 
-import com.taja.collector.domain.StationStatus;
+import com.taja.status.domain.StationStatus;
 import com.taja.station.domain.Station;
 import java.time.LocalDateTime;
 import reactor.core.publisher.Mono;
